requirements.txt
pandas==1.5.2
PyYAML==6.0
glob2
